+++ src/main/resources/db/migration/V16__Add_tcp_and_ping_checks.sql
@@ -1,0 +1,5 @@
+ALTER TABLE CheckEntity
+    ADD COLUMN host VARCHAR(255) NULL;
+
+ALTER TABLE CheckEntity
+    ADD COLUMN port INT NULL;